Start mic when program loads
There was an issue with the microphone only starting to record audio after a shush had been performed, perhaps because the mic was frozen until an audio playing activated it. Now mic.start() is called directly after mic is initialized instead of getting called in setup. This should fix the issue.