Feathers on the homescreen; Color scheme tweaks.
The background of arongil.com now has a 3D animation running in the background. It slows the page down for sure, but it finally fills all the white space that used to live on the main page.

Changed the nav bar on the main site and in /academy to a light orange for better contrast.