Feathers abide to deltaTime.
Previously, feathers would move depending on how often requestAnimationFrame called the main loop. On large screens for which THREE.js lagged, this meant slower movement.

Now, feathers move according to Date.now() so that their speed is the same on all screens, regardless of lag.

One issue was that users coming back to the page after a while would see the feathers all teleport super high then respawn down below, breaking the illusion of a continuous stream.

To fix this, any frames with deltaTimes higher than one second discard the movement of feathers. Users returning to the page won't notice the single frame that drops. This may be an issue on incredibly laggy machines that take longer than one second to load each frame. This is rare, however.